requirements.txt
pandas>=2.3.0
numpy>=2.3.0
scikit-learn>=1.7.0
xgboost>=3.1.0
joblib>=1.3.0
requests>=2.32.0
websocket-client>=1.9.0
python-dotenv>=1.0.1
loguru>=0.7.2
ta>=0.11.0
finta>=1.3
tvdatafeed==2.1.0
yfinance>=0.2.28
python-telegram-bot==13.15
pytz>=2024.1
# MetaTrader5 is optional (only available on Windows/Linux via vendor binary)
# MetaTrader5; sys_platform == "win32"
feedparser
requests
beautifulsoup4
vaderSentiment
transformers        # optional, heavy - only if you enable TF pipeline
torch               # required for many HF models - optional (heavy)
bs4
